requirements.txt
ipykernel
ipywidgets
python-dotenv
openai
langchain
llama-index
PyPDF2
docx2txt
